Enable support for presice ALU exceptions
Some architectures have a precise ALU exceptions reporting mode.  When
this mode is enabled, it ensures that if an ALU exception is delivered
to a wave, the wave's PC is at the instruction immediately following the
one that caused the exception to be raised.

This patch enables it in dbgapi, using precise memory operations as a
template.

Change-Id: I80e89bd2fde732501041cd0d0474cd5c4fb2161b


[ROCm/rocdbgapi commit: d2322002b813dd7c5278c3077200e3062d3caacf]