src/process: support attaching to a process with more than 16 queues
When attaching to a process with more than 16 queues, the current
version of dbgapi fails.  Here is what happens.

During the attach process, process_t::update_queues is called.  This
will call:

    os_driver ().queue_snapshot (snapshots.data (), snapshot_count,
		                 &queue_count,
                                 os_exception_mask_t::queue_new);

The default value for snapshot_count at this point is 16, so the driver
call will
  1) return exactly 16 entries (this is how many entries are
     available in snapshots), and
  2) clear the queue_new flag for all of those 16 queues.

Continuing its process, dbgapi will create a queue_t object for each of
those 16 entries, provision enough slots in `snapshots` to receive more
(all) queue descriptors, and it will come back to do a second call to
os_driver ().queue_snapshot ().

This second call will return the previous 16 entries plus all the
missing queue entries.  When processing the 16 first entries, we will
find a queue_t object (created during the first iteration), and then hit
this case:

    ... if ((os_queue_exception_status (queue_info)
            & os_exception_mask_t::queue_new)
            != os_exception_mask_t::none)
      { ... }
    else if (is_flag_set (flag_t::runtime_enable_during_attach))
      {
        dbgapi_assert (!queue && "no queue should exists before attach");
      }

Which means that when we are attaching, it is OK for queue_snapshot to
return unknown queues which do to have the queue_new flag set (this can
happen if another instance attached to the process and detached), but it
is not OK to return known queue entries without that flag set.  This
assertion is invalid when attaching to a process which have more than 16
queues, as discussed above.

This patch fixes that issue by removing that assertion.  Instead, if we
have a queue_t entry for that queue, we mark it as still active, and
carry on.

Change-Id: I726db7f18c8c90c85401881fcfb7811948aaa180


[ROCm/rocdbgapi commit: a74055fbd64171ea191dbefbfc9625237cddcfd4]