Allow building dbgapi as a static library
This allows building a static library, with:

 $ cmake -DBUILD_SHARED_LIBS=OFF

Building as shared library is still the default, so nothing changes if
you don't explicitly override BUILD_SHARED_LIBS.

BUILD_SHARED_LIBS is a predefined option.  As per
https://cmake.org/cmake/help/v3.30/command/add_library.html:

  "If no <type> is given the default is STATIC or SHARED based on the
  value of the BUILD_SHARED_LIBS variable."

So we just have to remove the explicit SHARED.  That's the easy part.

The hard part is adjusting the pkgconfig file to add the amd_comgr
dependency to Libs when static linking.  Or rather, find the right
bits to add to the pkgconfig file.

amd_comgr does not use pkgconfig, but it does provide cmake
find_package support.

For Windows, with amd_comgr built for configuration "Release", we see
in .../lc/lib/cmake/amd_comgr/amd_comgr-targets-release.cmake:

 # Import target "amd_comgr" for configuration "Release"
 set_property(TARGET amd_comgr APPEND PROPERTY IMPORTED_CONFIGURATIONS RELEASE)
 set_target_properties(amd_comgr PROPERTIES
   IMPORTED_IMPLIB_RELEASE "${_IMPORT_PREFIX}/lib/amd_comgr_2.lib"
   IMPORTED_LOCATION_RELEASE "${_IMPORT_PREFIX}/bin/amd_comgr_2.dll"
   )

For Linux, with amd_comgr built for configuration "RelWithDebInfo", we
see in
/opt/rocm/lib/cmake/amd_comgr/amd_comgr-targets-relwithdebinfo.cmake:

 # Import target "amd_comgr" for configuration "RelWithDebInfo"
 set_property(TARGET amd_comgr APPEND PROPERTY IMPORTED_CONFIGURATIONS RELWITHDEBINFO)
 set_target_properties(amd_comgr PROPERTIES
   IMPORTED_LOCATION_RELWITHDEBINFO "${_IMPORT_PREFIX}/lib/libamd_comgr.so.2.7.60200"
   IMPORTED_SONAME_RELWITHDEBINFO "libamd_comgr.so.2"
   )

So for Windows, we need to get at one of the
IMPORTED_IMPLIB_{RELEASE,RELWITHDEBINFO,DEBUG} properties, depending
on how amd_comgr was built, to find the right import library to use.
For Linux, I think we can assume that amd_comgr is built as a shared
library with an soname.  But we still have to get at
IMPORTED_LOCATION_{RELEASE,RELWITHDEBINFO,DEBUG}, to find the path to
use with -L<path> in the .pc file.

Unfortunately, cmake doesn't automatically create variables like
${amd_comgr_IMPORTED_IMPLIB} and ${amd_comgr_IMPORTED_LOCATION}
automatically depending on which amd_comgr configuration it decided to
use.  I wasn't able to find that information easily, which surprised
me.  So we have to determine them manually.  That is what the new
get_imported_property function does.

Here are examples of what the Libs line of the generated amd-dbgapi.pc
file looks like when building a static dbgapi library:

 Windows:

  Libs: -L${libdir} -lrocm-dbgapi -L/compute/drivers/compute/install/native/Release/x64/lc/lib -l:amd_comgr_2.lib

 Linux:

  Libs: -L${libdir} -lrocm-dbgapi -L/opt/rocm/lib -lamd_comgr

When dbgapi is built as a shared library (the default), we get the
same as before, on either Windows / Linux:

  Libs: -L${libdir} -lrocm-dbgapi

Change-Id: I486c5dac5055da5e0343c26a3c13ac393fa15409


[ROCm/rocdbgapi commit: d59eb7c86dfd4474da7cc85ac9c9cd0201aa9811]