Allow HSA headers to be found outside of hsa-runtime64
Dbgapi does not use the hsa-runtime, it just needs to find hsa headers
that contain definitions of structures used in the ABI used by the
runtime.

This patch makes it easier to build dbgapi on a system where the
hsa-runtime64 package in not available.  In such case, the user can
specify -DHSA_RUNTIME_INCLUDE_DIRECTORIES=/path/to/header when invoking
cmake so the build infrastructure knows where to find the required
headers.

The approach also allows the hsa headers to be found with the
CFLAGS="-I/path/to/hsa/headers/" approach (or imagine the hsa headers
are in /usr/include/).

Co-authored-by: Pedro.Alves <pedro.alves@amd.com>
Change-Id: I2f3f3aa1eb538e94dc0eded329254f9af4435def


[ROCm/rocdbgapi commit: 7f360fbf27bfadb624626267ed83fd97132e82bc]