CHANGELOG: Move gfx12 support in the 0.77.0 section
Rocm-6.2 ships with dbgapi-0.76.0.  Support for gfx12 architecture got
merged around the same time, under the same version number, but is not
part of the 6.2 release.  Version number should have been bumped when
merging this support.  As a result, things are confusing.

This patch updates the CHANGELOG file and moves the gfx12 support in
the 0.77.0 section, which will be the first released version to support
gfx12.

Change-Id: I2e3f17fc89147e6dc1acac9f7ad216db2d961347
Bug: SWDEV-485626


[ROCm/rocdbgapi commit: eec34f6dc3881ac4471b0479fdd7edaf76ae9c23]