Revert "[SWDEV-446215] Update cmake to put test libs in proper lib dir"
This reverts commit a33cdd7da67875b01c8f63bdf57396daf193542d.

Reason for revert: Because the gtest of amdsmi is different to other components so it was installed in a share/amdsmi/lib folder. It cannot be installed in a common folder such as /usr/local/bin or /usr/bin because all other components try to search those folder first.

 

This is breaking ROCmValidationSuite and other tools. Per Wang, Yanyao this should be reverted.

Change-Id: Id61bc6056fe41800e738616f39293e9b8762a377


[ROCm/amdsmi commit: b7789d46995ef2832b6771e371d8601f53e396d6]