gfx12_architecture_t::wave_set_halt: Fix setting halt
The gfx12 implementation changing the HALT bit.  Before this patch, the
implementation would flip bit 13 of WAVE_STATE_PRIV (which is bit 1 of
USER_PRIO) instead of bit 14.

As a result, when resuming a wave with exception (wave_t::set_state), we
would fail to keep the wave halted.  As a result, the wave could
continue executing and have unexpected side-effects.

This patch fixes this typo.

Bug: SWDEV-487950
Tested-by: Shahab Vahedi <shahab.vahedi@amd.com>
Change-Id: If96e4547cb3c0d19f0dfad5314ae7882be8ad4d3


[ROCm/rocdbgapi commit: 81a60b86e0f5cef1e609466272059799ce85108f]