set_exceptions: all known exceptions must be considered
The "convert_mask" lambda in "amdgcn_architecutrue_t::set_exceptions"
method is refactored in such a way that if an exception enum is not
considered, the compiler will bluntly let us know.  For instance:

 error: enumeration value 'overflow' not handled in switch [-Werror=switch]

Moreover, "mask" argument is not checked against every exception anymore.
Instead, only the set bits in "mask" are processed by means of a simple
"while" loop.

For the sake of consistency, the same sort of refactoring is made to:

- gfx940_t::set_exceptions ()
- gfx11_architecture_t::set_exceptions ()
- gfx12_architecture_t::set_exceptions ()

Co-Authored-By: Laurent Morichetti <laurent.morichetti@amd.com>
Change-Id: I47eaad563c74284a7a7686453ee3120b49b40e4e


[ROCm/rocdbgapi commit: c8bf1716d2d169d52f37a3d2a3722b41ad1feb87]