src/proces.cpp: Improve warning message for unsupported arch
When dbgapi detects a device which it cannot support, it produces a
message containing the device's name (found from the pci.ids file), but
does not provide the actual architecture name.

This patch changes this to print both the device's name and its
architecture (gfx number) in the warning message.

Change-Id: Ib5fb073b2fd61f7dd7033175c06c57c8374847ed


[ROCm/rocdbgapi commit: dd0833159b1966fba931e401755aac1bd4d3df5f]