Capture code object image during hsa_executable_freeze
It is possible for a client application to register a code object and
reuse its backing memory before the debug agent accesses it.  When that
happens, the code object is frozen, then copied to device memory and
executed.  When an exception is reported in a wave's execution, the
debug agent tries to load the code object, but at that moment the
backing memory has been modified, making it impossible for the debug
agent to load it.

This patch makes it so the debug agent captures an image of the code
object as soon as it detects it, so even if the client application
modifies the memory, the debug agent still has a valid copy it can use.

We use the already present hooks on hsa_executable_freeze and
hsa_executable_destroy to be notified on code object loading and
unloading.  When those methods are called, we notify the debug agent
worker thread to update the code object list.  This is handled by
DebugAgentWorker::update_code_object_list which encapsulates the
communication with the worker thread, and is blocking so we do not
return to the application before the code object inspection is
completed.

Bug: SWDEV-507543
Change-Id: I5fc362c43a5dbbe2595c47a85f03f23722bc4546


[ROCm/rocr_debug_agent commit: 00769879cfc4b85d9bbb7306a9cce8c1d68de4c5]