src/os_driver.cpp: Do not use std::basic_ostringstream<std::byte>
It is not correct to use a std::basic_ostringstream<std::byte> since it
would require std::char_traits to be defined for std::byte.  Such trait
is not defined for std::byte.

This issue causes build problem when against libc++-18 or above, as it
does not provide an implementation for std::char_traits<std::byte>.

This patch proposes to not rely on basic_ostringstream, but instead a
plain vector of std::byte.

Bug: https://github.com/ROCm/ROCdbgapi/issues/18
Bug: SWDEV-503716
Change-Id: I3deb0273ffdc6ba9977a6c29f25e41c12ffa7c64