Avoid name conflict for saved code objects
It is possible to see different code objects which have the same URI (if
they are loaded from the same location and have the same size).  If this
happens, the debug agent's -s option will dump a copy of the different
code objects in the same file.  This means that the last loaded code
object overrides the previous one(s).

This patch makes sure to use unique file names to save the code objects
by adding a "ID_" prefix to the filename, ID being the unique handle
ID provided by rocm-dbgapi.

Change-Id: I59e38cb0ad5b963a6a8d2223ee3cc0575ca37a24