Fix race in debug agent setup
There is a (mostly harmless) race issue when loading the debug agent.
When the runtime starts, it will at some point load the debug agent via
its OnLoad method.  This method will start a worker thread which will
handle all the dbgapi-based work.  However, the OnLoad method is not
waiting for this thread to complete its initial setup before it returns.

From there, 2 things can happen.

In the first scenario, the dbgapi thread gets all its initialization
before anything else has time to happen, and everything works as
expected.

In the second scenario, OnLoad can return before
amd_dbgapi_process_attach is called, and the runtime has time to start
to register a first code object.  In that case,
amd_dbgapi_process_attach will see the first code object, and so when we
eventually refresh the list in debug_agent_hsa_executable_freeze, we do
not see any change.  As a result, the debug agent does not refresh its
list of known code objects, and can miss one as reported by
SWDEV-516642.

There are 2 possible approaches to solve this problem.  The first one is
to have OnLoad wait for the dbgapi setup to complete before returning,
so we are sure to see the "new code object" event when it comes.  The
other approach would be to query the list of known code objects once
amd_dbgapi_process_attach returns.

In this patch, I go for the 1st approach as it seems slightly safer.
The second approach does rely on the assumption that nothing can happen
before loading a code object.  This seems a reasonable assumption, but
since option 1 does not rely on it, is seems a bit more suited.

Bug: SWDEV-516642
Change-Id: Ibb5b1124a2ca736402dcc408c45da69f8ba21963


[ROCm/rocr_debug_agent commit: 9132860d7e92872ce8eca32dcfd23961e292986d]