Don't use amd_queue_t::scratch_backing_memory_byte_size
A future change to the ROCR/CP ABI removes the
amd_queue_t::scratch_backing_memory_byte_size field which holds the
size, per XCC, of the scratch memory region allocated for the queue.

```
@@ -133,7 +133,7 @@ typedef struct AMD_QUEUE_ALIGN amd_queue_v2_s {
   uint32_t compute_tmpring_size;
   uint32_t scratch_resource_descriptor[4];
   uint64_t scratch_backing_memory_location;
-  uint64_t scratch_backing_memory_byte_size;
+  uint32_t reserved3[2];
   uint32_t scratch_wave64_lane_byte_size;
   amd_queue_properties32_t queue_properties;
   volatile uint64_t scratch_max_use_index;       /* V2 */
```

Since this information is also available in the tmpring_size register
(as tmpring_size.waves * tmpring_size.wavesize * BLOCK_SIZE), remove
the need for this field and rematerialize the scratch memory size when
computing the scratch memory region bounds.

When the ROCR is updated, the ROCR ABI version will be bumped to v11,
which this change is compatible with.

Tested on gfx942, gfx1030, and gfx1200.


[ROCm/rocdbgapi commit: b399310a2e450b859cf83535e32808c4c8de196c]