src/os_driver.cpp: Fix reading marketing name
The routine reading marketing name improperly uses
"std::basic_ios<CharT,Traits>::operator bool" to detect EOF.

This is an issue in this case because the routine tries to skip over
invalid lines (where DID/RID cannot be read properly).  When reaching
the end of the file, one cannot read DID, raising the "fail" bit.  From
there the procedure assumes the field is improperly formatted, clears
the bit, tries to skip until the end of the line (which does not change
the fail bit), and retries the loop.  On loop entry, the fail bit is not
set, so operator bool() returns true, and we retry.

Fix this by properly testing the EOF bit to ensure we exit the read loop
if no name has been found for the device.

Change-Id: Ic6c1342c4589b35d35bdad6f2ac8a4d65a81d41d