Prevent VGPR_DEALLOC in est/vector_add_memory_fault.cpp
A recent change in the dirver has disabled the precise memory for
gfx11/gfx12.  Some tests for the debug agent to rely on precise memory
to catch some accents when `s_sendmsg msg(MSG_DEALLOC_VGPRS)` is used.
As we cannot use precise memory anymore, to use the optnone attribute to
prevent the compiler from releasing the VGPRs.

Change-Id: I53b0d9c35ba40116e3e911ea5b5fe2b903d5f021
Bug: SWDEV-524985