src/os_driver.h: Improve how invalid or error queues are reported
KFD does encode the ERROR and INVALID state of a queue inside the
queue_id.  This patch proposes to update how dbgapi handles those bits
internally.  Instead of carrying them around in the queue_id as KFD
does, we split them into separate fields.

This change has an implication for queue_t::os_queue_id which now should
only be called for valid queues (either in the running or suspended
state) as otherwise we could return an os_queue_id which has been
re-used by the system for another queue.  Only one place was calling
this method for a potentially invalid queue, but as this was only for
logging purposse.

Change-Id: I8416f0cd2d9bb0a45b730419838e7064dbc2dd85