queue_t::get_info: Fix assertion failure when accessing the OS_ID
Having a client asking for the AMD_DBGAPI_QUEUE_INFO_OS_ID request to
amd_dbgapi_queue_get_info currently leads to an assertion failure.

This is because of this piece of code:

    std::optional<amd_dbgapi_os_queue_id_t> os_id;
    dbgapi_assert (os_id.has_value ());

The os_id variable is missing a proper initialisation by calling
queue_t::os_queue_id ().

This commit fixes this.

Bug: SWDEV-537602
Change-Id: If1d47608a6cc7a0014b83cadf73dca1b449b29af