Tests: Fix race in sigquit test
The SIGQUIT test relies on hard-coded delays to decide when to send the
SIGQUIT to the process, and expects some output.  On some configuration,
this test can fail, especially when running the test with
HIP_ENABLE_DEFERRED_LOADING=0.  On systems with many GPUs, loading all
available code objects on each GPU can take time.  By the time the
SIGQUIT arrives, no wave has been created, so the debug agent's output
is empty, leading to test failure.

This patch tries to remove the timeout.  To do so, it uses a sentinel
variable accessible by both the host and device.  When the device starts
executing the kernel, it writes to the sentinel variable.  The host can
see the value changed, and write to its stdout to notify the test
framework that the wave has started execution.  Only after this message
has been seen, the test send SIGQUIT, and checks for the output
produced.

Bug: SWDEV-548768
Change-Id: Ib178d04934c286abaed55bd81c5601a098ac227f