Enable stack traces when compiling with Clang
libbacktrace is a C API library that is used to print stack traces
when a fatal exception is raised.  It is is normally installed with
GCC within its private tree (/usr/lib/gcc/...), and is not found by
other compilers unless the search path for includes and libraries
is explicitely set to point to that location.

Use the --print-file-name compiler option (supported by both gcc and
clang) to find the location where libbacktrace.a is installed, and
deduce the libbacktrace install prefix from that location.

Change-Id: Ib9ed89d481f6a88c1e8f11de4eb082de69e08aab


[ROCm/rocdbgapi commit: a1dfdb013a74b8011cb27fcf98c581d4f263258e]