Allow dbgapi to have internal address spaces
Currently, all the address spaces tied to an architecture are returned
to the debugger when doing the
amd_dbgapi_architecture_address_space_list call.  There are cases where
we might want to keep address spaces internal (not expose them to the
client) as there is nothing meaningful the client can do with those
address spaces.

This patch proposes to add a `is_internal` predicate to address_space_t
(currently checking if the address space has an associated dwarf number,
but this could be changed in the future if needed), and use this
predicate to filter the address-spaces returned to the client.

Change-Id: I81cd3d44e05fffb194bdca044738fe9b0a2790bd


[ROCm/rocdbgapi commit: 053d4df0dcccb7cb8eca484adaf4ed9865c2d5d4]