Drop libva-amdgpu use for RHEL8 (#660)
After further inspection, it looks like libva is new enough for decode
support, so we've dropped the requirement on the mesa end. Some encode
features are missing on < 2.16, but as per our current RHEL8 support
policy, these features are no longer supported.

Ubuntu 22.04 is still supported to have a full encode feature set, so we
can't drop the libva requirement quite yet.

SWDEV-548213

Signed-off-by: Jeremy Newton <Jeremy.Newton@amd.com>
Co-authored-by: Kiriti Gowda <kiritigowda@gmail.com>