process_exited_exception_t takes a process_id
In a future patch, we will have cases where we can throw
process_exited_exception_t from a const method of a process_t.  However,
the process_exited_exception_t ctor takes a non-const reference to a
process, which won't be available.

To avoid this issue, this patch changes process_exited_exception_t to
accept a amd_dbgapi_process_id_t argument, which will later be used to
retrieve a non-const reference to the process_t object when needed.

Change-Id: I59f9642713e9737ad87382654a853f8448bf5a6e