Add explicit methods to read host memory
Previous patches use global and host memory interchangeably, they are
indistinguishable on Linux, but this is not the case for all platforms.
This patch tries to disambiguate.

In order to have proper exception informations for failed host read
operations, this patch also introduces a global host_address_space
class, and a singleton of it.

Change-Id: I4af588303a3e69c9f04b87f964f10f04385a0106


[ROCm/rocdbgapi commit: 327132914b99762e221a7c7afafa689fe304a2b6]