src/queue: Use proper type to hold return of os_queue_id
In queue_t::get_info, we have:

    case AMD_DBGAPI_QUEUE_INFO_OS_ID:
      {
        std::optional<amd_dbgapi_os_queue_id_t> os_id = os_queue_id ();

however, os_queue_id () is defined as:

    std::optional<os_queue_id_t> os_queue_id () const;

with

    /* From include/amd-dbgapi.h.in.  */
    typedef uint64_t amd_dbgapi_os_queue_id_t;
    /* From src/os_driver.h.  */
    using os_queue_id_t = uint32_t

So things do not line up, and this is confusing.

This works because std::optional<T> has a constructor such as:

    template< class U >
    optional( const optional<U>& other );

which is valid if std::is_constructible_v<T, const U&> is true.

Although this works, I find is highly confusing.  This patch proposes to
make things clearer by not relying on that particular implicit constructor.

Change-Id: If24c9da9985c75633ce3407f89d3799536081833


[ROCm/rocdbgapi commit: 77e77972ce1e15dfc67eb2b3b71a9eb53a3635d2]