Turn AIS_WARN_UNSAFE_BUFFER_OPS on by default (#90)
This affects clang/llvm C++ only. We currently raise no warnings with
this switched on. If it becomes problematic in the future, we can switch
it back off, but it's useful to turn on to encourage developers to find
other solutions if this raises warnings.

This is only turned on by default w/ clang. For gcc, it remains off
since it's not useful. It will now emit a FATAL_ERROR if enabled with
gcc.

[ROCm/hipFile commit: bd15dbbddd93110f180bdf3b6f617e40c2c3e0e0]