CI: Update container names (#93)
This isn't an issue on isolated runners, but on a host with multiple
runners could induce a conflict in Docker with trying to start a
container with the same name as an existing container. This would only
occur on matrix jobs since the matrix parameter is not actually added to
the `job_id`. I don't believe we ever hit this issue however since this
was introduced as we switched away from self-hosted machines.

[ROCm/hipFile commit: f498d4c8053fc8a6b67d97b06778417434517524]