Add a sleep-on-exit argument that can be used on system and unit tests. (#124)
When using bpftrace, symbols are looked up asynchronously in certain
cases. The program needs to be running for this symbol lookup to work,
so a short running program may exit before symbols are known for calls
of interest. This option allows you to add a sleep on exit to give time
for symbol lookup to work.

This won't help with destructors of objects with static lifetimes, but
will generally let you trace calls of interest.

[ROCm/hipFile commit: 7a1cf64cf4d0cdbc6ddf10d9ea68f57cf6be25c5]